реализовал связь между списком предложений и картой
