+++ src/components/place-card-list/place-card-list.tsx
@@ -1,25 +1,16 @@
 import { PreviewOffer } from '../../types/offer';
-import {useState} from 'react';
 import PlaceCard from '../place-card/place-card';
 import { PlaceCardType } from '../../const';
 
 type PlaceCardListProps = {
   offers: PreviewOffer[];
+  callback?: React.MouseEventHandler<HTMLElement>;
 }
 
-function PlaceCardList({offers} : PlaceCardListProps) {
-  const [activeCard, setActiveCard] = useState(' ');
-
-  const handlePlaceCardMouseOver = (evt : React.MouseEvent) => {
-    evt.preventDefault();
-    const id = evt.currentTarget.getAttribute('data-id');
-    if (id !== null) {
-      setActiveCard(id);
-    }
-  };
+function PlaceCardList({offers, callback} : PlaceCardListProps) {
 
   function getPlaceCards(cards: PreviewOffer[]) {
-    return Array.from({length: offers.length}, (_, index : number) => <PlaceCard callback={handlePlaceCardMouseOver} offer={cards[index]} type={PlaceCardType.City} key={cards[index].id}/>);
+    return Array.from({length: offers.length}, (_, index : number) => <PlaceCard callback={callback} offer={cards[index]} type={PlaceCardType.City} key={cards[index].id}/>);
   }
 
   return (
